fix path

+++ install/install.sh
@@ -66,7 +66,20 @@ MAIN()
   EZBIN=/usr/local/bin/ezinstall_patch_mgr
   if [ ! -h $EZBIN ]; then
     echo "=> File doesn't exist"
-    ln -s /usr/local/ezyadmin/connector_patch_manager/script/ezinstall_patch_mgr /usr/local/bin/ezinstall_patch_mgr
+
+    # default file connector_patch_manager-latest
+    FILE=/usr/local/ezyadmin/connector_patch_manager-latest/script/ezinstall_patch_mgr
+    if test -f "$FILE"; then
+        ln -s /usr/local/ezyadmin/connector_patch_manager-latest/script/ezinstall_patch_mgr /usr/local/bin/ezinstall_patch_mgr
+    fi
+
+    # dir connector_patch_manager
+    FILE=/usr/local/ezyadmin/connector_patch_manager/script/ezinstall_patch_mgr
+    if test -f "$FILE"; then
+        ln -s /usr/local/ezyadmin/connector_patch_manager/script/ezinstall_patch_mgr /usr/local/bin/ezinstall_patch_mgr
+    fi
+
+    # ln -s /usr/local/ezyadmin/connector_patch_manager/script/ezinstall_patch_mgr /usr/local/bin/ezinstall_patch_mgr
   fi
   echo "$DATE_NOW" > /usr/local/ezyadmin/connector_patch_manager/install.log
 }
